start button works but needs to be refactored and lives text is now on the screen
