+++ index.html
@@ -8,9 +8,12 @@
 </head>
 <body>
   <canvas id = "myCanvas" width = "900" height = "600" style="border: 1px solid black;"></canvas>
+  <div id = "score-box">
+    <p id = "score-number">Score: 0</p>
+  </div>
   <div id="start-container">
-  <button id="start-button" type="button">Start Game</button>
-</div>
+    <button id="start-button" type="button">Start Game</button>
+  </div>
   <script src="main.bundle.js"></script>
 </body>
 </html>
