+++ lib/Ball.js
@@ -9,7 +9,7 @@ class Ball {
   }
 
   draw(context) {
-    context.fillStyle = "#E400D0";
+    context.fillStyle = "#00F6FB";
     context.beginPath();
     context.arc(this.x, this.y, 10, 0, Math.PI * 2);
     context.fill();
